Reusing format for PA and Coverage
